finishedButNotTested

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: aichida <aichida@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/28 15:26:14 by aichida           #+#    #+#             */
-/*   Updated: 2024/04/28 15:35:53 by aichida          ###   ########.fr       */
+/*   Updated: 2024/05/02 14:38:52 by aichida          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while ((i < n) && !(s1[i] == '\0' && s2[i] == '\0'))
